need to refactor with liri instead of spotifyTest
