+++ log.txt
@@ -1,1 +1,2 @@
 node spotifyTest.js spotify-this-song "I Want it That Way"
+node spotifyTest.js spotify-this-song "I Want it That Way"
