+++ spotify-module.js
@@ -10,14 +10,11 @@ var GET = function () {
       if (err) {
         return console.log(`Error occurred: ${err}`);
       }
-      // console.log(data.tracks.items[1]);
       console.log(`artists: ${data.tracks.items[1].album.artists[0].name}`);
       console.log(`preview url: ${data.tracks.items[1].preview_url}`);
       console.log(`external url: ${data.tracks.items[1].external_urls.spotify}`);
       console.log(`album name: ${data.tracks.items[1].album.name}`);
     });
-    // console.log(err.config);
   };
 };
-// console.log(GetSong('over the hills and far away'));
 module.exports = GET;
